requirements.txt
jedi==0.19.2
parso==0.8.5
pip == 25.3
pyside6==6.10.1
git+https://github.com/nate-maxwell/PySide6TK.git
setuptools==80.9.0
shiboken6==6.10.1
gitpython
